+++ ts/features/idpay/onboarding/machine/machine.ts
@@ -35,7 +35,8 @@ export const idPayOnboardingMachine = setup({
     navigateToInitiativeMonitoringScreen: notImplementedStub,
     closeOnboarding: notImplementedStub,
     handleSessionExpired: notImplementedStub,
-    navigateToInputFormScreen: notImplementedStub
+    navigateToInputFormScreen: notImplementedStub,
+    navigateToEnableNotificationScreen: notImplementedStub
   },
   actors: {
     getInitiativeInfo: fromPromise<InitiativeDataDTO, string>(
@@ -86,7 +87,9 @@ export const idPayOnboardingMachine = setup({
       context.activeTextConsentPage >=
       getInputFormSelfDeclarationFromContext(context).length - 1,
     isSessionExpired: ({ event }: { event: IdPayOnboardingEvents }) =>
-      "error" in event && event.error === InitiativeFailureType.SESSION_EXPIRED
+      "error" in event && event.error === InitiativeFailureType.SESSION_EXPIRED,
+    shouldShowEnableNotificationOnClose: ({ context }) =>
+      !context.isPushNotificationsEnabled
   }
 }).createMachine({
   id: "idpay-onboarding",
@@ -104,7 +107,8 @@ export const idPayOnboardingMachine = setup({
         "start-onboarding": {
           guard: "assertServiceId",
           actions: assign(({ event }) => ({
-            serviceId: event.serviceId
+            serviceId: event.serviceId,
+            currentStep: 1
           })),
           target: "LoadingInitiative"
         }
@@ -249,14 +253,23 @@ export const idPayOnboardingMachine = setup({
       on: {
         next: [
           {
+            actions: assign(({ context }) => ({
+              currentStep: context.currentStep + 1
+            })),
             guard: "hasSelfDecalrationList",
             target: "DisplayingSelfDeclarationList"
           },
           {
+            actions: assign(({ context }) => ({
+              currentStep: context.currentStep + 1
+            })),
             target: "AcceptingCriteria"
           }
         ],
         back: {
+          actions: assign(({ context }) => ({
+            currentStep: context.currentStep - 1
+          })),
           target: "#idpay-onboarding.DisplayingInitiativeInfo"
         }
       }
@@ -288,10 +301,16 @@ export const idPayOnboardingMachine = setup({
           on: {
             back: [
               {
+                actions: assign(({ context }) => ({
+                  currentStep: context.currentStep - 1
+                })),
                 guard: "hasPdndCriteria",
                 target: "#idpay-onboarding.DisplayingPdndCriteria"
               },
               {
+                actions: assign(({ context }) => ({
+                  currentStep: context.currentStep - 1
+                })),
                 target: "#idpay-onboarding.DisplayingInitiativeInfo"
               }
             ],
@@ -306,7 +325,10 @@ export const idPayOnboardingMachine = setup({
             next: [
               {
                 guard: "hasMultiSelfDeclarationList",
-                target: "DisplayingMultiSelfDeclarationList"
+                target: "DisplayingMultiSelfDeclarationList",
+                actions: assign(({ context }) => ({
+                  currentStep: context.currentStep + 1
+                }))
               },
               {
                 guard: "hasInputFormDeclaration",
@@ -327,6 +349,7 @@ export const idPayOnboardingMachine = setup({
               on: {
                 "select-multi-consent": {
                   actions: assign(({ context, event }) => ({
+                    currentStep: context.currentStep + 1,
                     selfDeclarationsMultiAnwsers: {
                       ...context.selfDeclarationsMultiAnwsers,
                       [context.selfDeclarationsMultiPage]: event.data
@@ -341,20 +364,29 @@ export const idPayOnboardingMachine = setup({
                       "hasBooleanSelfDeclarationList"
                     ]),
                     target:
-                      "#idpay-onboarding.DisplayingSelfDeclarationList.DisplayingBooleanSelfDeclarationList"
+                      "#idpay-onboarding.DisplayingSelfDeclarationList.DisplayingBooleanSelfDeclarationList",
+                    actions: assign(({ context }) => ({
+                      selfDeclarationsMultiPage:
+                        +context.selfDeclarationsMultiPage - 1,
+                      currentStep: context.currentStep - 1
+                    }))
                   },
                   {
                     guard: and(["isFirstMultiConsentPage", "hasPdndCriteria"]),
-                    target: "#idpay-onboarding.DisplayingPdndCriteria"
+                    target: "#idpay-onboarding.DisplayingPdndCriteria",
+                    actions: assign(({ context }) => ({
+                      selfDeclarationsMultiPage:
+                        +context.selfDeclarationsMultiPage - 1,
+                      currentStep: context.currentStep - 1
+                    }))
                   },
                   {
                     guard: "isFirstMultiConsentPage",
-                    target: "#idpay-onboarding.DisplayingInitiativeInfo"
-                  },
-                  {
+                    target: "#idpay-onboarding.DisplayingInitiativeInfo",
                     actions: assign(({ context }) => ({
                       selfDeclarationsMultiPage:
-                        +context.selfDeclarationsMultiPage - 1
+                        +context.selfDeclarationsMultiPage - 1,
+                      currentStep: context.currentStep - 1
                     }))
                   }
                 ]
@@ -374,7 +406,8 @@ export const idPayOnboardingMachine = setup({
                 {
                   actions: assign(({ context }) => ({
                     selfDeclarationsMultiPage:
-                      +context.selfDeclarationsMultiPage + 1
+                      +context.selfDeclarationsMultiPage + 1,
+                    currentStep: context.currentStep + 1
                   })),
                   target: "DisplayingMultiSelfDeclarationItem"
                 }
@@ -394,7 +427,8 @@ export const idPayOnboardingMachine = setup({
                     selfDeclarationsTextAnswers: {
                       ...context.selfDeclarationsTextAnswers,
                       [context.activeTextConsentPage]: event.criteria
-                    }
+                    },
+                    currentStep: context.currentStep + 1
                   })),
                   target: "EvaluatingInputForm"
                 },
@@ -425,7 +459,8 @@ export const idPayOnboardingMachine = setup({
                   },
                   {
                     actions: assign(({ context }) => ({
-                      activeTextConsentPage: +context.activeTextConsentPage - 1
+                      activeTextConsentPage: +context.activeTextConsentPage - 1,
+                      currentStep: context.currentStep - 1
                     }))
                   }
                 ]
@@ -473,8 +508,43 @@ export const idPayOnboardingMachine = setup({
       }
     },
 
+    EnableNotification: {
+      entry: "navigateToEnableNotificationScreen",
+      on: {
+        "update-notification-status": {
+          actions: assign(({ event }) => ({
+            isPushNotificationsEnabled: event.isPushNotificationEnabled
+          }))
+        },
+        back: {
+          actions: "closeOnboarding",
+          target: "Idle"
+        },
+        close: {
+          actions: "closeOnboarding",
+          target: "Idle"
+        }
+      }
+    },
+
     OnboardingCompleted: {
-      entry: "navigateToCompletionScreen"
+      entry: "navigateToCompletionScreen",
+      on: {
+        "update-notification-status": {
+          actions: assign(({ event }) => ({
+            isPushNotificationsEnabled: event.isPushNotificationEnabled
+          }))
+        },
+        close: [
+          {
+            target: "EnableNotification",
+            guard: "shouldShowEnableNotificationOnClose"
+          },
+          {
+            actions: "closeOnboarding"
+          }
+        ]
+      }
     },
 
     OnboardingFailure: {