+++ ts/features/itwallet/lifecycle/saga/__tests__/checkWalletInstanceStateSaga.test.ts
@@ -21,13 +21,18 @@ jest.mock("@pagopa/io-react-native-crypto", () => ({
 }));
 
 describe("checkWalletInstanceStateSaga", () => {
+  const mockPid = {
+    credentialType: "PersonIdentificationData",
+    credentialId: "dc_sd_jwt_PersonIdentificationData",
+    format: "dc+sd-jwt"
+  } as StoredCredential;
   // TODO: improve the mocked store's typing, do not use DeepPartial
   it("Does not check the wallet state when the wallet is INSTALLED", () => {
     const store: DeepPartial<GlobalState> = {
       features: {
         itWallet: {
           issuance: { integrityKeyTag: O.none },
-          credentials: { eid: O.none, credentials: [] }
+          credentials: { credentials: {} }
         }
       }
     };
@@ -47,7 +52,7 @@ describe("checkWalletInstanceStateSaga", () => {
             integrityServiceStatus: "ready",
             integrityKeyTag: O.some("aac6e82a-e27e-4293-9b55-94a9fab22763")
           },
-          credentials: { eid: O.none, credentials: [] },
+          credentials: { credentials: {} },
           environment: {
             env: "prod"
           }
@@ -76,7 +81,7 @@ describe("checkWalletInstanceStateSaga", () => {
           issuance: {
             integrityKeyTag: O.some("aac6e82a-e27e-4293-9b55-94a9fab22763")
           },
-          credentials: { eid: O.none, credentials: [] },
+          credentials: { credentials: {} },
           environment: {
             env: "prod"
           }
@@ -104,7 +109,9 @@ describe("checkWalletInstanceStateSaga", () => {
           issuance: {
             integrityKeyTag: O.some("3396d31e-ac6a-4357-8083-cb5d3cda4d74")
           },
-          credentials: { eid: O.some({} as StoredCredential), credentials: [] },
+          credentials: {
+            credentials: { [mockPid.credentialId]: mockPid }
+          },
           environment: {
             env: "prod"
           }
@@ -132,7 +139,9 @@ describe("checkWalletInstanceStateSaga", () => {
           issuance: {
             integrityKeyTag: O.some("3396d31e-ac6a-4357-8083-cb5d3cda4d74")
           },
-          credentials: { eid: O.some({} as StoredCredential), credentials: [] },
+          credentials: {
+            credentials: { [mockPid.credentialId]: mockPid }
+          },
           environment: {
             env: "prod"
           }
@@ -158,7 +167,9 @@ describe("checkWalletInstanceStateSaga", () => {
       features: {
         itWallet: {
           issuance: { integrityKeyTag: O.none },
-          credentials: { eid: O.some({} as StoredCredential), credentials: [] }
+          credentials: {
+            credentials: { [mockPid.credentialId]: mockPid }
+          }
         }
       }
     };