+++ ts/features/settings/common/screens/__tests__/ProfileMainScreen.test.tsx
@@ -21,7 +21,7 @@ const renderComponent = () => {
   const globalState = appReducer(undefined, applicationChangeState("active"));
   const store = createStore(
     appReducer,
-    globalState as PreloadedState<ReturnType<typeof appReducer>>
+    globalState as unknown as PreloadedState<ReturnType<typeof appReducer>>
   );
 
   return {